major restructuring and clean up
