+++ ext/BeamTrackingBeamlinesExt/BeamTrackingBeamlinesExt.jl
@@ -2,9 +2,9 @@ module BeamTrackingBeamlinesExt
 using Beamlines, BeamTracking, GTPSA, StaticArrays, KernelAbstractions, AtomicAndPhysicalConstants
 using Beamlines: isactive, deval, unsafe_getparams, isnullspecies
 using BeamTracking: get_N_particle, R_to_beta_gamma, R_to_gamma, R_to_pc, R_to_v, beta_gamma_to_v,
-                    @makekernel, Coords, KernelCall, KernelChain, push, TimeDependentParam, RefState, launch!
+                    @makekernel, Coords, KernelCall, KernelChain, push, TimeDependentParam, RefState, 
+                    launch!, AbstractYoshida
 import BeamTracking: track!
-
 
 include("utils.jl")
 
@@ -37,13 +37,9 @@ function track!(
   return bunch
 end
 
-
-include("aperture.jl")
-include("alignment.jl")
 include("unpack.jl")
 include("scibmadstandard.jl")
-include("linear.jl")
 include("exact.jl")
-include("integration.jl")
+include("yoshida.jl")
 
 end
