--- ext/BeamTrackingBeamlinesExt/alignment.jl
@@ -1,35 +1,0 @@
-@inline function alignment(tm, bunch, alignmentparams, bendparams, L, entering::Bool)
-  if !isactive(alignmentparams); return nothing; end
-
-  x_off = alignmentparams.x_offset
-  y_off = alignmentparams.y_offset
-  z_off = alignmentparams.z_offset
-  x_rot = alignmentparams.x_rot
-  y_rot = alignmentparams.y_rot
-  tilt  = alignmentparams.tilt
-
-  ele_orient = 1   ## Future work: Need to extend this for reversed elements.
-
-  #
-
-  if isactive(bendparams) && (bendparams.g_ref != 0 || bendparams.tilt_ref != 0)
-    if entering
-      dr, q = BeamTracking.coord_alignment_bend_entering(x_off, y_off, z_off, 
-                x_rot, y_rot, tilt, bendparams.g_ref, bendparams.tilt_ref, ele_orient, L)
-      return KernelCall(BeamTracking.track_coord_transform!, (dr, q))
-    else
-      dr, q = BeamTracking.coord_alignment_bend_exiting(x_off, y_off, z_off, 
-                x_rot, y_rot, tilt, bendparams.g_ref, bendparams.tilt_ref, ele_orient, L)
-      return KernelCall(BeamTracking.track_coord_transform!, (dr, q))
-    end
-
-  else
-    if entering
-      return KernelCall(BeamTracking.track_alignment_straight_entering!, (x_off, y_off, z_off, 
-                                                     x_rot, y_rot, tilt, ele_orient, L))
-    else
-      return KernelCall(BeamTracking.track_alignment_straight_exiting!, (x_off, y_off, z_off, 
-                                                     x_rot, y_rot, tilt, ele_orient, L))
-    end
-  end
-end
