+++ ext/BeamTrackingBeamlinesExt/utils.jl
@@ -150,7 +150,7 @@ end
 
 function rf_omega(rfparams, circumference, species, R_ref)
   if rfparams.harmon_master
-    tilde_m, gamsqr_0, beta_0 = ExactTracking.drift_params(species, R_ref)
+    tilde_m, gamsqr_0, beta_0 = BeamTracking.drift_params(species, R_ref)
     return 2*pi*rfparams.harmon*C_LIGHT*beta_0/circumference
   else
     return 2*pi*rfparams.rf_frequency
