+++ src/kernels/alignment.jl
@@ -3,9 +3,9 @@
   q = inv_rot_quaternion(x_rot, y_rot, tilt)
   L2 = 0.5 * L * ele_orient
 
-  track_translation!(i, coords, (x_off, y_off, 0.0), 0.0)
-  dz_new = track_rotation!(i, coords, q, -L2 - z_off)
-  track_isochronous_drift!(i, coords, -dz_new - L2)
+  translation!(i, coords, (x_off, y_off, 0.0), 0.0)
+  dz_new = rotation!(i, coords, q, -L2 - z_off)
+  isochronous_drift!(i, coords, -dz_new - L2)
 end
 
 #
@@ -18,13 +18,13 @@ end
   L2 = 0.5 * L * ele_orient
 
   q = rot_quaternion(x_rot, y_rot, tilt)
-  dz_new = track_rotation!(i, coords, q, L2)
+  dz_new = rotation!(i, coords, q, L2)
 
   alive = (coords.state[i] == STATE_ALIVE)
   v[i,XI] = vifelse(alive, v[i,XI] + x_off, v[i,XI])
   v[i,YI] = vifelse(alive, v[i,YI] + y_off, v[i,YI])
 
-  track_isochronous_drift!(i, coords, L2 - z_off - dz_new)
+  isochronous_drift!(i, coords, L2 - z_off - dz_new)
 end
 
 #---------------------------------------------------------------------------------------------------
@@ -40,10 +40,10 @@ inverse of this transformation.
 """
 
 @makekernel fastgtpsa=true function track_coord_transform!(i, coords::Coords, r, q)
-  track_translation!(i, coords, (r[1], r[2], 0.0), 0.0)
-  dz_new = track_rotation!(i, coords, quat_inv(q), -r[3])
+  translation!(i, coords, (r[1], r[2], 0.0), 0.0)
+  dz_new = rotation!(i, coords, quat_inv(q), -r[3])
   ##println("***AA: $(coords.v[1,:]) :: $dz_new")
-  track_isochronous_drift!(i, coords, -dz_new)
+  isochronous_drift!(i, coords, -dz_new)
   ##println("***BB: $(coords.v[1,:])")
 end
