+++ src/kernels/aperture.jl
