+++ src/kernels/transforms.jl
@@ -1,5 +1,5 @@
 """
-    track_isochronous_drift!(i, coords, L)
+    isochronous_drift!(i, coords, L)
 
 Isochronous drifting used in the alignment algorithm. 
 The difference between this and the standard drift is that the reference time does not change
@@ -8,65 +8,56 @@ as the particle drifts from beginning to end.
 ## Arguments
 - `L`:       element length, in meters
 """
-@inline function track_isochronous_drift!(i, coords::Coords, L)
-  @FastGTPSA begin @inbounds begin 
-    v = coords.v
+@makekernel fastgtpsa=true function isochronous_drift!(i, coords::Coords, L)
+  v = coords.v
 
-    rel_p = 1 + v[i,PZI]
-    P_t2 = v[i,PXI]*v[i,PXI] + v[i,PYI]*v[i,PYI]
-    P_s2 = rel_p*rel_p - P_t2
+  rel_p = 1 + v[i,PZI]
+  P_t2 = v[i,PXI]*v[i,PXI] + v[i,PYI]*v[i,PYI]
+  P_s2 = rel_p*rel_p - P_t2
 
-    good_momenta = (P_s2 > 0)
-    alive = (coords.state[i] == STATE_ALIVE)
-    coords.state[i] = vifelse(!good_momenta & alive, STATE_LOST, coords.state[i])
-    alive = (coords.state[i] == STATE_ALIVE)
+  good_momenta = (P_s2 > 0)
+  alive = (coords.state[i] == STATE_ALIVE)
+  coords.state[i] = vifelse(!good_momenta & alive, STATE_LOST, coords.state[i])
+  alive = (coords.state[i] == STATE_ALIVE)
 
-    P_s2_1 = one(P_s2)
-    P_s = sqrt(vifelse(good_momenta, P_s2, P_s2_1))
+  P_s2_1 = one(P_s2)
+  P_s = sqrt(vifelse(good_momenta, P_s2, P_s2_1))
 
-    new_x = v[i,XI] + v[i,PXI] * L / P_s
-    new_y = v[i,YI] + v[i,PYI] * L / P_s
-    new_z = v[i,ZI] - rel_p * L / P_s
+  new_x = v[i,XI] + v[i,PXI] * L / P_s
+  new_y = v[i,YI] + v[i,PYI] * L / P_s
+  new_z = v[i,ZI] - rel_p * L / P_s
 
-    v[i,XI] = vifelse(alive, new_x, v[i,XI])
-    v[i,YI] = vifelse(alive, new_y, v[i,YI])
-    v[i,ZI] = vifelse(alive, new_z, v[i,ZI])
-  end end
+  v[i,XI] = vifelse(alive, new_x, v[i,XI])
+  v[i,YI] = vifelse(alive, new_y, v[i,YI])
+  v[i,ZI] = vifelse(alive, new_z, v[i,ZI])
 end
 
 #---------------------------------------------------------------------------------------------------
 """
-    function track_translation!(i, coords::Coords, dr, z0) -> z1
+    function translation!(i, coords::Coords, dr, z0) -> z1
 
 Transform phase-space particle coordinates when the coordinate system is translated by the vector `dr`.
 
 ## Arguments
 - `dr`:      coordinate system translation.
 - `z0`       The particle longitudinal distance from the center of rotation. 
-
-## Output
-- `z1`        The particle longitudinal distance from the center of rotation after rotation. 
-
 """
-@inline function track_translation!(i, coords::Coords, dr, z0)
-  @FastGTPSA begin @inbounds begin 
-    v = coords.v
-    alive = (coords.state[i] == STATE_ALIVE)
+@makekernel fastgtpsa=true function translation!(i, coords::Coords, dr, z0)
+  v = coords.v
+  alive = (coords.state[i] == STATE_ALIVE)
 
-    x = v[i,XI] - dr[1]
-    y = v[i,YI] - dr[2]
-    z = z0    - dr[3]
+  x = v[i,XI] - dr[1]
+  y = v[i,YI] - dr[2]
+  z = z0    - dr[3]
 
-    v[i,XI] = vifelse(alive, x, v[i,XI])
-    v[i,YI] = vifelse(alive, y, v[i,YI])
-    return z
-  end end
+  v[i,XI] = vifelse(alive, x, v[i,XI])
+  v[i,YI] = vifelse(alive, y, v[i,YI])
 end
 
 #---------------------------------------------------------------------------------------------------
 
 """
-    track_rotation!(i, coords::Coords, q_inv, z_0) -> z_1
+    rotation!(i, coords::Coords, q_inv, z_0) -> z_1
 
 Particle coordinate rotation. This rotates both `(x, y)` and `(px, py, pz)` phase space coordinates.
 
@@ -79,8 +70,7 @@ Particle coordinate rotation. This rotates both `(x, y)` and `(px, py, pz)` phas
 
 Returned is the new longitudinal offset from the center of rotation.
 """
-
-@inline function track_rotation!(i, coords::Coords, q_inv, z_0) 
+@inline function rotation!(i, coords::Coords, q_inv, z_0) 
   @FastGTPSA begin @inbounds begin
     v = coords.v
     rel_p = 1 + v[i,PZI]
