+++ test/IntegrationTracking_test.jl
@@ -1,4 +1,4 @@
-@testset "SplitIntegration" begin
+@testset "Yoshida" begin
   @testset "Kernels" begin
     function multipole_args(::Type{T}) where {T}
       ms = SA[1, 6]
@@ -80,7 +80,7 @@
       tilde_m = mc2/p0c
       beta_0 = 1/sqrt(1 + tilde_m^2)
       params = (0, 0, false, tilde_m, beta_0, a, 0, w, w_inv, Kn0, mm, kn, ks)
-      ker = IntegrationTracking.bkb_multipole!
+      ker = BeamTracking.bkb_multipole!
       num_steps = 10
       ds_step = T(0.2)
       return ker, params, nothing, ds_step, num_steps, 0, 0, L
@@ -102,7 +102,7 @@
       gamsqr_0 = 1 + 1/tilde_m^2
       beta_0 = 1/sqrt(1 + tilde_m^2)
       params = (0, 0, false, beta_0, gamsqr_0, tilde_m, a, mm, kn, ks)
-      ker = IntegrationTracking.dkd_multipole!
+      ker = BeamTracking.dkd_multipole!
       num_steps = 1
       ds_step = T(2)
       return ker, params, nothing, ds_step, num_steps, 0, 0, L
@@ -127,27 +127,27 @@
     end
     
     # Scalar parameters
-    test_map("bmad_maps/thin_dipole.jl", KernelCall(ExactTracking.multipole_kick!, multipole_args(Float64)); tol=1e-14, no_scalar_allocs=true)
-    test_map("bmad_maps/sol_quad.jl", KernelCall(IntegrationTracking.sks_multipole!, sk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
-    test_map("bmad_maps/skew_quad_mk.jl", KernelCall(IntegrationTracking.mkm_quadrupole!, mk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
-    test_map("bmad_maps/sex_dec.jl", KernelCall(IntegrationTracking.dkd_multipole!, dk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
-    test_map("bmad_maps/sex_dec.jl", KernelCall(IntegrationTracking.order_two_integrator!, integrator_args(Float64)); tol=5e-10, no_scalar_allocs=true)
-    test_map("bmad_maps/order_four.jl", KernelCall(IntegrationTracking.order_four_integrator!, integrator_args(Float64)); tol=5e-10, no_scalar_allocs=true)
-    test_map("bmad_maps/order_six.jl", KernelCall(IntegrationTracking.order_six_integrator!, integrator_args(Float64)); tol=5e-9, no_scalar_allocs=true)
-    test_map("bmad_maps/order_eight.jl", KernelCall(IntegrationTracking.order_eight_integrator!, integrator_args(Float64)); tol=5e-9, no_scalar_allocs=true)
-    test_map("bmad_maps/straight_dipole_bk.jl", KernelCall(IntegrationTracking.order_six_integrator!, bk_straight_args(Float64)); tol=2e-6, no_scalar_allocs=true)
-    test_map("bmad_maps/pure_rf.jl", KernelCall(IntegrationTracking.cavity!, cavity_args(Float64)); tol=2e-7, no_scalar_allocs=true)
+    test_map("bmad_maps/thin_dipole.jl", KernelCall(BeamTracking.multipole_kick!, multipole_args(Float64)); tol=1e-14, no_scalar_allocs=true)
+    test_map("bmad_maps/sol_quad.jl", KernelCall(BeamTracking.sks_multipole!, sk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
+    test_map("bmad_maps/skew_quad_mk.jl", KernelCall(BeamTracking.mkm_quadrupole!, mk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
+    test_map("bmad_maps/sex_dec.jl", KernelCall(BeamTracking.dkd_multipole!, dk_args(Float64)); tol=5e-10, no_scalar_allocs=true)
+    test_map("bmad_maps/sex_dec.jl", KernelCall(BeamTracking.order_two_integrator!, integrator_args(Float64)); tol=5e-10, no_scalar_allocs=true)
+    test_map("bmad_maps/order_four.jl", KernelCall(BeamTracking.order_four_integrator!, integrator_args(Float64)); tol=5e-10, no_scalar_allocs=true)
+    test_map("bmad_maps/order_six.jl", KernelCall(BeamTracking.order_six_integrator!, integrator_args(Float64)); tol=5e-9, no_scalar_allocs=true)
+    test_map("bmad_maps/order_eight.jl", KernelCall(BeamTracking.order_eight_integrator!, integrator_args(Float64)); tol=5e-9, no_scalar_allocs=true)
+    test_map("bmad_maps/straight_dipole_bk.jl", KernelCall(BeamTracking.order_six_integrator!, bk_straight_args(Float64)); tol=2e-6, no_scalar_allocs=true)
+    test_map("bmad_maps/pure_rf.jl", KernelCall(BeamTracking.cavity!, cavity_args(Float64)); tol=2e-7, no_scalar_allocs=true)
 
     # GTPSA parameters
-    test_map("bmad_maps/thin_dipole.jl", KernelCall(ExactTracking.multipole_kick!, multipole_args(TPS64{D10})); tol=1e-14)
-    test_map("bmad_maps/sol_quad.jl", KernelCall(IntegrationTracking.sks_multipole!, sk_args(TPS64{D10})); tol=5e-10)
-    test_map("bmad_maps/skew_quad_mk.jl", KernelCall(IntegrationTracking.mkm_quadrupole!, mk_args(TPS64{D10})); tol=5e-10)
-    test_map("bmad_maps/sex_dec.jl", KernelCall(IntegrationTracking.dkd_multipole!, dk_args(TPS64{D10})); tol=5e-10)
-    test_map("bmad_maps/sex_dec.jl", KernelCall(IntegrationTracking.order_two_integrator!, integrator_args(TPS64{D10})); tol=5e-10)
-    test_map("bmad_maps/order_four.jl", KernelCall(IntegrationTracking.order_four_integrator!, integrator_args(TPS64{D10})); tol=5e-10)
-    test_map("bmad_maps/order_six.jl", KernelCall(IntegrationTracking.order_six_integrator!, integrator_args(TPS64{D10})); tol=5e-9)
-    test_map("bmad_maps/order_eight.jl", KernelCall(IntegrationTracking.order_eight_integrator!, integrator_args(TPS64{D10})); tol=5e-9)
-    test_map("bmad_maps/straight_dipole_bk.jl", KernelCall(IntegrationTracking.order_six_integrator!, bk_straight_args(TPS64{D10})); tol=2e-6)
-    test_map("bmad_maps/pure_rf.jl", KernelCall(IntegrationTracking.cavity!, cavity_args(TPS64{D10})); tol=2e-7)
+    test_map("bmad_maps/thin_dipole.jl", KernelCall(BeamTracking.multipole_kick!, multipole_args(TPS64{D10})); tol=1e-14)
+    test_map("bmad_maps/sol_quad.jl", KernelCall(BeamTracking.sks_multipole!, sk_args(TPS64{D10})); tol=5e-10)
+    test_map("bmad_maps/skew_quad_mk.jl", KernelCall(BeamTracking.mkm_quadrupole!, mk_args(TPS64{D10})); tol=5e-10)
+    test_map("bmad_maps/sex_dec.jl", KernelCall(BeamTracking.dkd_multipole!, dk_args(TPS64{D10})); tol=5e-10)
+    test_map("bmad_maps/sex_dec.jl", KernelCall(BeamTracking.order_two_integrator!, integrator_args(TPS64{D10})); tol=5e-10)
+    test_map("bmad_maps/order_four.jl", KernelCall(BeamTracking.order_four_integrator!, integrator_args(TPS64{D10})); tol=5e-10)
+    test_map("bmad_maps/order_six.jl", KernelCall(BeamTracking.order_six_integrator!, integrator_args(TPS64{D10})); tol=5e-9)
+    test_map("bmad_maps/order_eight.jl", KernelCall(BeamTracking.order_eight_integrator!, integrator_args(TPS64{D10})); tol=5e-9)
+    test_map("bmad_maps/straight_dipole_bk.jl", KernelCall(BeamTracking.order_six_integrator!, bk_straight_args(TPS64{D10})); tol=2e-6)
+    test_map("bmad_maps/pure_rf.jl", KernelCall(BeamTracking.cavity!, cavity_args(TPS64{D10})); tol=2e-7)
   end
 end
